Add execute permission

+++ applications/bash/configuration/bashrc/configure.sh
@@ -8,7 +8,9 @@ location_path="$HOME/.bashrc"
 directory_path="$(dirname "${location_path}")"
 file_name="$(basename "${location_path}")"
 
-ls -la "${location_path}"
+if [[ -f "${location_path}" ]];
+  chmod +x "${location_path}" || fail "Unable to add execute permissions on file ${location_path}"
+fi
 
 mkdir -p "${directory_path}" || fail "Failed to create directory ${directory_path}"
 cp -n "${content_path}" "${location_path}" || fail "Failed to copy content from ${content_path} to ${location_path}"
